Fix the URL for get registrations and get it to work
Use event_id parameter instead of id parameter in the registrations URL and
fix the path of the URL for oauth calls. Also, fetch information about the
team and the player, when fetching a registration, instead of just ids.

Modified by @punchagan to remove the function to get help for an end-point,
and some debug/commented code.