Add view to display event registrations
The view currently displays all the registrations for an event,
grouped by team name. A player can have multiple registrations, each
for a different role. Such players are listed only once, along with
all the roles.

For making the development of the event_page view faster, currently
the calls to get event_data are outside the event_page method, and
hard-coded for a specific event. Once a caching strategy for this
data is figured out, this code should be moved into the event_page
method.

Co-authored-by: Puneeth Chaganti <punchagan@muse-amuse.in>