Use a file based cache for caching API responses from UC
We make API calls to Ultimate Central to fetch the tournament information,
and registration information. Ideally, we'd like to have updated information
whenever someone is viewing pages displaying this information, but we do not
want to make too many API calls. So, we cache these responses for 15 minutes
before making new requests for this data.

In order not to write our own caching code, and to be able to re-use the
Django caching infrastructure, this commit adds a couple of un-exposed views
which return the json data from the API response. These views are cached
views, that Django manages caching for.

Using a file based cache seems good enough for our current purposes.

Closes #2

Co-authored-by: Ranjan Khyadad <ranjankhyadad@gmail.com>
Co-authored-by: Pavan Nayak <pavan.nayak272@gmail.com>