Forms for existing models was incorrectly being marked as unchanged
The logic in `has_changed` was buggy and only marking a form as changed if
the `date` and `wfdf_userid` have been changed. But, for existing
accreditation data in the DB, accreditation type, wfdf_userid, date etc can
be changed independent of each other. This changed data was not incorrectly
being ignored, because of the buggy `has_changed`.

This commit fixes that by checking if we are saving the form for an instance
already in the DB, in which case we mark the form as changed, when any field
has changed. Also, the fields `date` and `wfdf_userid` are no longer allowed
to be `blank`, while still allowing the user to fill in data for only some
players in a team, thanks to `has_changed` ignoring forms that don't have
`date` and `wfdf_userid` filled in.