Improve the logic to ignore forms as unchanged
Previously, we ignored all forms as unchanged if both the date and the
wfdf_userid were not filled in. But, unchanged forms are not validated, and
are silently ignored. This leads to a bad UX.

This commit marks any form that has any of the forms with type, date or
wfdf_userid changed as a changed form. This means any form with incomplete
data is validated and the user is prompted to fill in missing data, rather
than silently ignoring such forms.