Add player accreditation information to event page
We do a single query to the DB to fetch accreditation information for all
the registered players for an event. We then group the registrations by
team, and add the accreditation information for each player to this
regrouped data.

The event page now displays accreditation status against each user in the
event page, if there is a valid accreditation in the DB. Accreditations are
considered valid if the date is newer than 18 months ago. We use arrow to
compute the oldest valid date for accreditations.

Co-authored-by: Ranjan Khyadad <ranjankhyadad@gmail.com>