Fix a bug where re-using a formset causes problems saving it
When a formset is successfully saved, some of it's internal metadata about
initial number of forms, etc., doesn't get updated. This causes a bug when
trying to re-save the formset with new data. The formset is being re-used
because the POST response redirects to the same page.

Thanks to A2 for the bug report.