Increase the cache timeout to an year
We are fetching and updating data only for new events. We also switched the
app views to never talk to the UC API. This means the data for the older
events is always fetched from the cache, and it would be nice to be able to
view the event pages of older events for some time after the event has
finished.