Add help-text to the date field in accreditation model
Also, copied and modified the `table_inline_formset.html` and
`help_text.html` from crispy forms to override where the help text is
displayed. We display it in the header of the table, instead of displaying
it in every row. `help_text.html` is set to an empty file to prevent the
help text from showing up in every row of the table form.

Closes #22